proxy/requirements.txt
fastapi
uvicorn
sqlparse
mysql-connector-python
